+++ pyoxidizer/src/starlark/file_resource.rs
@@ -8,7 +8,7 @@ use {
     super::python_resource::PythonExtensionModuleFlavor,
     super::python_resource::{
         PythonBytecodeModule, PythonExtensionModule, PythonPackageDistributionResource,
-        PythonResourceData, PythonSourceModule,
+        PythonPackageResource, PythonSourceModule,
     },
     super::target::{BuildContext, BuildTarget, ResolvedTarget, RunMode},
     super::util::{
@@ -243,8 +243,8 @@ impl FileManifest {
 
                 Ok(())
             }
-            "PythonResourceData" => {
-                let m = resource.downcast_apply(|m: &PythonResourceData| m.data.clone());
+            "PythonPackageResource" => {
+                let m = resource.downcast_apply(|m: &PythonPackageResource| m.data.clone());
                 warn!(logger, "adding resource file {} to {}", m.full_name, prefix);
                 m.add_to_file_manifest(&mut self.manifest, &prefix)
                     .or_else(|e| {
@@ -552,7 +552,9 @@ mod tests {
         super::super::testutil::*,
         super::*,
         crate::py_packaging::resource::DataLocation,
-        crate::py_packaging::resource::{PythonModuleSource, PythonPackageResource},
+        crate::py_packaging::resource::{
+            PythonModuleSource, PythonPackageResource as RawPackageResource,
+        },
         std::path::PathBuf,
     };
 
@@ -620,8 +622,8 @@ mod tests {
             manifest: RawFileManifest::default(),
         });
 
-        let v = Value::new(PythonResourceData {
-            data: PythonPackageResource {
+        let v = Value::new(PythonPackageResource {
+            data: RawPackageResource {
                 full_name: "foo/bar/resource.txt".to_string(),
                 leaf_package: "foo.bar".to_string(),
                 relative_name: "resource.txt".to_string(),